chore: refactored data-grid-component and load-more-button
